backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
python-dotenv==1.0.0
selenium==4.15.2
lxml==4.9.3
apscheduler==3.10.4